Protocol "mode" is now bitwise based on bit-packing

+++ SerialX-juss/src/main/java/org/ugp/serialx/juss/protocols/UniversalObjectInstantiationProtocol.java
@@ -92,7 +92,7 @@ public class UniversalObjectInstantiationProtocol<T> extends SerializationProtoc
 	}
 	
 	@Override
-	public byte getMode() 
+	public long getMode() 
 	{
 		return MODE_DESERIALIZE;
 	}
